Formatting changes to align with vec_ps_refl
Code has not yet been altered, but simply
formatted to better align with the style present
in Vec_ps_refl.cc and Vec_ps_refl.h
